refactor(application): 결과값 출력 파라미터 수정
