+++ src/main/java/racingcar/application/NameSeparator.java
@@ -1,4 +1,4 @@
-package racingcar.racingapplication;
+package racingcar.application;
 
 import java.util.List;
