+++ src/main/java/racingcar/application/Printer.java
@@ -1,13 +1,13 @@
-package racingcar.racingapplication;
+package racingcar.application;
 
 import java.util.List;
-import racingcar.domain.Cars;
+import racingcar.domain.Result;
 
 public interface Printer {
 
     void print(String message);
 
-    void printAfterGameResult(List<Cars> afterGameCars);
+    void printAfterGameResult(Result result);
 
     void printFinalWinner(List<String> winner);
 }
