+++ src/main/java/racingcar/application/RacingApplication.java
@@ -1,4 +1,4 @@
-package racingcar.racingapplication;
+package racingcar.application;
 
 import java.util.ArrayList;
 import java.util.List;
