+++ src/main/java/racingcar/configuration/AppConfig.java
@@ -2,9 +2,9 @@ package racingcar.configuration;
 
 import racingcar.nameseparator.OriginSeparator;
 import racingcar.printer.PrintRacingResult;
-import racingcar.racingapplication.NameSeparator;
-import racingcar.racingapplication.Printer;
-import racingcar.racingapplication.Reader;
+import racingcar.application.NameSeparator;
+import racingcar.application.Printer;
+import racingcar.application.Reader;
 import racingcar.reader.ConsoleReader;
 
 public class AppConfig {
