+++ src/main/java/racingcar/printer/PrintRacingResult.java
@@ -2,7 +2,8 @@ package racingcar.printer;
 
 import java.util.List;
 import racingcar.domain.Cars;
-import racingcar.racingapplication.Printer;
+import racingcar.domain.Result;
+import racingcar.application.Printer;
 
 public class PrintRacingResult implements Printer {
 
@@ -12,9 +13,9 @@ public class PrintRacingResult implements Printer {
     }
 
     @Override
-    public void printAfterGameResult(List<Cars> afterGameCars) {
+    public void printAfterGameResult(Result afterGameCars) {
         System.out.println("실행 결과");
-        for (Cars cars : afterGameCars) {
+        for (Cars cars : afterGameCars.finalResultCars()) {
             System.out.println(cars.makeResultOutput());
         }
     }
